# Models/Dtos/CreateCustomerDto.cs
namespace HighThroughputApi.Dtos
{
    public class CreateCustomerDto
    {
        public int Id { get; set; } 
        public string FirstName { get; set; } = null!;
        public string LastName { get; set; } = null!;
        public string Email { get; set; } = null!;

    }
}