# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "HighThroughputApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/v{version}/Customer/{id}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateCustomerDto"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "If-Match",
            "description": "Required ETag header for concurrency control",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/Customer": {
      "post": {
        "tags": [
          "Customer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCustomerDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateCustomerDto"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Customer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CreateCustomerDto"
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Orders/customer/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "If-Match",
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "If-Match",
            "description": "Required ETag header for concurrency control",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateOrderDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "If-Match",
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "If-Match",
            "description": "Required ETag header for concurrency control",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/Orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "If-Match",
            "type": "string"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateOrderDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "CreateCustomerDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateOrderDto": {
      "type": "object",
      "properties": {
        "customerId": {
          "format": "int32",
          "type": "integer"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItemDto"
          }
        }
      },
      "additionalProperties": false
    },
    "Customer": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "rowVersion": {
          "format": "byte",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Item": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "stock": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "rowVersion": {
          "format": "byte",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "customerId": {
          "format": "int32",
          "type": "integer"
        },
        "customer": {
          "$ref": "#/definitions/Customer"
        },
        "rowVersion": {
          "format": "byte",
          "type": "string"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        }
      },
      "additionalProperties": false
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "$ref": "#/definitions/Order"
        },
        "itemId": {
          "format": "int32",
          "type": "integer"
        },
        "item": {
          "$ref": "#/definitions/Item"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "rowVersion": {
          "format": "byte",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OrderItemDto": {
      "type": "object",
      "properties": {
        "itemId": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "UpdateOrderDto": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItemDto"
          }
        }
      },
      "additionalProperties": false
    }
  }
}